Fix: moved base inside types
