+++ src/types/class.ts
@@ -23,7 +23,7 @@
  * @module class
  */
 
-import Base from "../base";
+import Base from "./base";
 import { getInstance } from "../ibmi";
 import { getColumns, generateDetailTable, checkProcedureExists, checkViewExists, checkTableFunctionExists } from "../tools";
 import * as vscode from 'vscode';
