+++ src/types/jobDescription.ts
@@ -16,7 +16,7 @@
  * @module jobdescription
  */
 
-import Base from "../base";
+import Base from "./base";
 import { getInstance } from "../ibmi";
 import { getColumns, generateDetailTable, checkViewExists, executeSqlIfExists } from "../tools";
 import * as vscode from 'vscode';
