+++ src/types/jobQueue.ts
@@ -14,7 +14,7 @@
  * @module jobqueue
  */
 
-import Base from "../base";
+import Base from "./base";
 import { IBMiObject, CommandResult } from '@halcyontech/vscode-ibmi-types';
 import { Components } from "../webviewToolkit";
 import { getInstance } from "../ibmi";
