+++ src/types/messageFile.ts
@@ -15,7 +15,7 @@
  * @module messagefile
  */
 
-import Base from "../base";
+import Base from "./base";
 import { IBMiObject, CommandResult } from '@halcyontech/vscode-ibmi-types';
 import { getInstance } from "../ibmi";
 import { Tools } from '@halcyontech/vscode-ibmi-types/api/Tools';
