+++ src/types/outputQueue.ts
@@ -16,7 +16,7 @@
  * @module outputqueue
  */
 
-import Base from "../base";
+import Base from "./base";
 import { IBMiObject, CommandResult } from '@halcyontech/vscode-ibmi-types';
 import { Components } from "../webviewToolkit";
 import { getInstance } from "../ibmi";
